introduced mutation to update waves query data with cached wave
