+++ src/providers/queries/index.ts
@@ -30,4 +30,3 @@ export * from './pointQueries';
 export * from './postQueries';
 export * from './walletQueries';
 export * from './leaderboardQueries';
-export * from './wavesQueries';
