+++ src/providers/queries/postQueries/index.ts
@@ -1,3 +1,4 @@
 import * as postQueries from './postQueries';
+import * as wavesQueries from './wavesQueries';
 
-export { postQueries };
+export { postQueries, wavesQueries };
