--- src/providers/queries/wavesQueries/index.ts
@@ -1,3 +1,0 @@
-import * as wavesQueries from './wavesQueries';
-
-export { wavesQueries };
